Move geo response classes to query-metric-api, ensure json/xml/html responses and map link work from webservice
