+++ warehouse/query-core/src/main/java/datawave/query/jexl/visitors/GeoFeatureVisitor.java
@@ -12,12 +12,12 @@ import org.apache.log4j.Logger;
 import org.geotools.geojson.geom.GeometryJSON;
 import org.locationtech.jts.io.WKTReader;
 
+import datawave.microservice.querymetric.QueryGeometry;
 import datawave.query.jexl.functions.GeoFunctionsDescriptor;
 import datawave.query.jexl.functions.GeoWaveFunctionsDescriptor;
 import datawave.query.jexl.functions.JexlFunctionArgumentDescriptorFactory;
 import datawave.query.jexl.functions.arguments.JexlArgumentDescriptor;
 import datawave.webservice.common.logging.ThreadConfigurableLogger;
-import datawave.webservice.query.map.QueryGeometry;
 
 /**
  * This visitor will traverse the query tree, and extract both the geo function and associated query geometry (as GeoJSON).
