+++ warehouse/query-core/src/test/java/datawave/query/jexl/visitors/GeoFeatureVisitorTest.java
@@ -9,8 +9,8 @@ import java.util.Set;
 import org.apache.commons.jexl2.parser.JexlNode;
 import org.junit.Test;
 
+import datawave.microservice.querymetric.QueryGeometry;
 import datawave.query.jexl.JexlASTHelper;
-import datawave.webservice.query.map.QueryGeometry;
 
 public class GeoFeatureVisitorTest {
     private static final String GEO_FIELD_QUERY = "(((GEO_FIELD >= '0208' && GEO_FIELD <= '020d') || GEO_FIELD == '0202') && (geowave:intersects(GEO_FIELD, 'POLYGON((10 10, -10 10, -10 -10, 10 -10, 10 10 ))')))";
