+++ web-services/query/src/main/java/datawave/webservice/query/map/QueryGeometryHandler.java
@@ -3,6 +3,7 @@ package datawave.webservice.query.map;
 import java.util.List;
 
 import datawave.microservice.querymetric.BaseQueryMetric;
+import datawave.microservice.querymetric.QueryGeometryResponse;
 
 public interface QueryGeometryHandler {
